requirements.txt
torch>=2.0.0
numpy>=1.21.0
gymnasium>=0.29.0
PyYAML>=6.0.1
imageio>=2.31.0
pathlib>=1.0.1
typing-extensions>=4.7.0
